Exceptions added

+++ src/main/java/com/example/demo/web/controller/EmployeeController.java
@@ -2,6 +2,7 @@ package com.example.demo.web.controller;
 
 import com.example.demo.business.repository.EmployeeRepository;
 import com.example.demo.business.repository.model.EmployeeDAO;
+import com.example.demo.handlers.ExceptionHandler;
 import com.example.demo.model.Employee;
 import com.example.demo.service.EmployeeService;
 import com.sun.istack.NotNull;
@@ -41,7 +42,7 @@ public class EmployeeController {
     public ResponseEntity<List<EmployeeDAO>> findAllEmployees() {
         List<EmployeeDAO> employeeList = employeeService.getAllEmployees();
         if (employeeList.isEmpty()) {
-            return ResponseEntity.notFound().build();
+            throw new ExceptionHandler("List is empty");
         }
         return ResponseEntity.ok(employeeList);
     }
@@ -49,18 +50,17 @@ public class EmployeeController {
     @GetMapping("/{id}")
     public ResponseEntity<EmployeeDAO> findEmployeeById(@ApiParam(value = "Employee id", required = true) @NotNull @PathVariable Long id) {
         Optional<EmployeeDAO> employee = employeeService.getEmployeeById(id);
-        if (!employee.isPresent()) {
-            log.warn("Employee with id {} not found.", id);
-            return ResponseEntity.notFound().build();
+        if (!employee.isPresent() || (id<0)) {
+            throw new ExceptionHandler("Employee with id not found - " + id);
         }
         return employee.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
     }
 
     @PostMapping("/add")
     public ResponseEntity<Employee> saveEmployee(@Valid @RequestBody EmployeeDAO employee, BindingResult bindingResult) throws Exception {
-        if (bindingResult.hasErrors()) {
-            log.error("Employee not created: error {}", bindingResult);
-        }
+            if(bindingResult.hasErrors()){
+                throw new ExceptionHandler("Invalid request body");
+            }
 
         return new ResponseEntity<>(employeeService.saveEmployee(employee), HttpStatus.CREATED);
 
@@ -70,26 +70,24 @@ public class EmployeeController {
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<Void> deleteEmployeeById(@PathVariable Long id) {
         Optional<EmployeeDAO> employee = employeeService.getEmployeeById(id);
-        if (!(employee.isPresent())) {
-            log.warn("Employee for delete with id {} is not found.", id);
-            return ResponseEntity.notFound().build();
+        if (!(employee.isPresent()||(id<0))) {
+            throw new ExceptionHandler("Employee with id not found - " + id);
         }
         employeeService.deleteEmployeeById(id);
-        log.debug("Employee with id {} is deleted: {}", id, employee);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 
     }
 
     @PutMapping("/update/{id}")
-    public ResponseEntity<Employee> updateEmployee(@PathVariable Long id, @RequestBody EmployeeDAO employee, BindingResult bindingResult) {
+    public ResponseEntity<Employee> updateEmployee(@PathVariable Long id,@Valid @RequestBody EmployeeDAO employee, BindingResult bindingResult) {
         if (bindingResult.hasErrors() || !id.equals(employee.getId())) {
-            log.warn("Employee for update with id {} not found", id);
-            return ResponseEntity.notFound().build();
+            throw new ExceptionHandler("Invalid request body");
         }
         employeeService.saveEmployee(employee);
         return new ResponseEntity<>(HttpStatus.CREATED);
 
     }
+
 
 
 }
