+++ src/main/java/com/example/demo/web/controller/JobController.java
@@ -2,6 +2,7 @@ package com.example.demo.web.controller;
 
 import com.example.demo.business.repository.JobRepository;
 import com.example.demo.business.repository.model.JobDAO;
+import com.example.demo.handlers.ExceptionHandler;
 import com.example.demo.service.JobService;
 import com.sun.istack.NotNull;
 import lombok.extern.log4j.Log4j2;
@@ -16,6 +17,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.validation.Valid;
 import java.util.List;
 import java.util.Optional;
 
@@ -32,37 +34,30 @@ public class JobController {
     public ResponseEntity<List<JobDAO>> findAllJobs() {
         List<JobDAO> jobDAOList = jobService.getAllJobs();
         if (jobDAOList.isEmpty()) {
-            return ResponseEntity.notFound().build();
-        }
-        return ResponseEntity.ok(jobDAOList);
-    }
+            throw new ExceptionHandler("list is empty");}
+        return ResponseEntity.ok(jobDAOList);}
 
     @GetMapping("/{id}")
     public ResponseEntity<JobDAO> findJobById(@NotNull @PathVariable Long id) {
         Optional<JobDAO> job = jobService.getJobById(id);
-        if (!job.isPresent()) {
-            log.warn("Job with id {} not found.", id);
-        } else {
-            log.debug("Job with id {} is found: {}", id, job);
-        }
+        if (!job.isPresent() || id < 0) {
+            throw new ExceptionHandler("Job with id not found - " + id);}
         return job.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
     }
 
     @PostMapping("/add")
-    public ResponseEntity<JobDAO> saveJob(@RequestBody JobDAO jobDAO) {
+    public ResponseEntity<JobDAO> saveJob(@Valid @RequestBody JobDAO jobDAO) {
         JobDAO jobSaved = jobService.saveJob(jobDAO);
         return new ResponseEntity<>(jobSaved, HttpStatus.CREATED);
     }
+
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<Void> deleteJobId(@PathVariable Long id) {
         Optional<JobDAO> jobDAO = jobService.getJobById(id);
-        if (!(jobDAO.isPresent())) {
-            log.warn("Job for delete with id {} is not found.", id);
-            return ResponseEntity.notFound().build();
-        }
+        if (!(jobDAO.isPresent()) || id < 0) {
+            throw new ExceptionHandler("No job with id found" + id);}
         jobService.deleteJobById(id);
-        log.debug("Job with id {} is deleted: {}", id, jobDAO);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-
     }
+
 }
